safari-600.7-branch/LayoutTests/svg/dom/SVGTransformList-expected.txt

This test checks the SVGTransformList API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check passing invalid arguments to 'createSVGTransformFromMatrix'
PASS transform.createSVGTransformFromMatrix() threw exception TypeError: Not enough arguments.
PASS transform.createSVGTransformFromMatrix(svgElement.createSVGTransform()) threw exception TypeError: Argument 1 ('matrix') to SVGTransformList.createSVGTransformFromMatrix must be an instance of SVGMatrix.
PASS transform.createSVGTransformFromMatrix(svgElement) threw exception TypeError: Argument 1 ('matrix') to SVGTransformList.createSVGTransformFromMatrix must be an instance of SVGMatrix.
PASS transform.createSVGTransformFromMatrix('aString') threw exception TypeError: Argument 1 ('matrix') to SVGTransformList.createSVGTransformFromMatrix must be an instance of SVGMatrix.
PASS transform.createSVGTransformFromMatrix(1) threw exception TypeError: Argument 1 ('matrix') to SVGTransformList.createSVGTransformFromMatrix must be an instance of SVGMatrix.
PASS transform.createSVGTransformFromMatrix(true) threw exception TypeError: Argument 1 ('matrix') to SVGTransformList.createSVGTransformFromMatrix must be an instance of SVGMatrix.
PASS transform.createSVGTransformFromMatrix(undefined) threw exception Error: TypeMismatchError: DOM Exception 17.
PASS transform.createSVGTransformFromMatrix(null) threw exception Error: TypeMismatchError: DOM Exception 17.
PASS successfullyParsed is true

TEST COMPLETE
